requirements.txt
apache-airflow-providers-amazon==8.27.0